# backend/server.js
// Error handlers for uncaught exceptions
process.on('uncaughtException', (err) => {
    console.error('FATAL: Uncaught Exception:', err);
    process.exit(1);
});

process.on('unhandledRejection', (reason, promise) => {
    console.error('FATAL: Unhandled Rejection at:', promise, 'reason:', reason);
    process.exit(1);
});

const express = require('express');
const cors = require('cors');
const axios = require('axios');

const swaggerUi = require('swagger-ui-express');
const YAML = require('yamljs');
const path = require('path');
const fs = require('fs');

const app = express();
const port = 3001;

const getBranchesData = () => {
    try {
        const branchesPath = path.resolve(__dirname, '..', 'frontend', 'src', 'constants', 'branches.data.json');
        const raw = fs.readFileSync(branchesPath, 'utf-8');
        const parsed = JSON.parse(raw);
        return Array.isArray(parsed) ? parsed : [];
    } catch (err) {
        console.error('⚠ Failed to load branches data:', err.message);
        return [];
    }
};

const BRANCHES_DATA = getBranchesData();

const ensurePrefixed = (code) => (code && code.startsWith('I') ? code : `I${code}`);

// Load OpenAPI spec with error handling
let openApiSpec = {};
try {
    const openApiPath = path.join(__dirname, 'openapi.yaml');
    console.log('Loading OpenAPI spec from:', openApiPath);
    openApiSpec = YAML.load(openApiPath) || {};
    console.log('✓ OpenAPI spec loaded');
} catch (err) {
    console.error('⚠ Failed to load OpenAPI spec:', err.message);
}

// Serve Swagger UI at /api-docs
app.use('/api-docs', swaggerUi.serve, swaggerUi.setup(openApiSpec));

app.use(cors());

// Import database - it will auto-initialize
console.log('Importing database module...');
const { db, initDatabase } = require('./database.js');
console.log('✓ Database module imported');
app.use(express.json());

// Error handling middleware for invalid JSON
app.use((err, req, res, next) => {
    if (err instanceof SyntaxError && err.status === 400 && 'body' in err) {
        console.error('Bad JSON:', err);
        return res.status(400).json({ error: 'Invalid JSON' });
    }
    next();
});

const inferDeviceType = (port) => {
    const p = Number(port);
    if (p === 8888) return 'windows11';
    if (p === 8081) return 'android';
    return 'unknown';
};

const normalizePath = (pathname, search, hash) => {
    const pathPart = pathname || '';
    const searchPart = search || '';
    const hashPart = hash || '';
    return `${pathPart}${searchPart}${hashPart}` || '/';
};

const buildCachedUrl = (device) => {
    if (device.last_url) return device.last_url;
    if (device.ip && device.port) {
        const path = device.path || '';
        const normalizedPath = path.startsWith('/') || path === '' ? path : `/${path}`;
        return `http://${device.ip}:${device.port}${normalizedPath}`;
    }
    return null;
};

const normalizeBranchCode = (branch) => {
    if (!branch) return '';
    if (branch.startsWith('I') && branch.length > 1) return branch.slice(1);
    return branch;
};

const buildBranchCandidates = (branch) => {
    const normalized = normalizeBranchCode(branch);
    const prefixed = normalized ? `I${normalized}` : '';
    const candidates = [branch, normalized, prefixed]
        .filter(Boolean)
        .map((code) => code.trim());
    return [...new Set(candidates)];
};

const updateDeviceNetworkInfo = (deviceName, info) => new Promise((resolve, reject) => {
    const { ip, port, path, finalUrl, deviceType } = info;
    db.run(
        `UPDATE devices SET ip = ?, port = ?, path = ?, last_url = ?, device_type = ? WHERE name = ?`,
        [ip, port, path, finalUrl, deviceType, deviceName],
        (err) => {
            if (err) return reject(err);
            resolve();
        }
    );
});

const resolveViaNetwork = async (deviceName, branchCode) => {
    const branchCandidates = buildBranchCandidates(branchCode);
    let lastStatus = null;
    let lastUrl = null;

    for (const candidate of branchCandidates) {
        const publicUrl = `http://qr.studiobox.vn:9096/qr/${candidate}/${deviceName}`;
        lastUrl = publicUrl;
        const response = await axios.get(publicUrl, {
            maxRedirects: 0,
            validateStatus: null,
            timeout: 5000,
        });
        lastStatus = response.status;

        if (response.status >= 300 && response.status < 400 && response.headers.location) {
            const finalUrl = response.headers.location;
            const parsed = new URL(finalUrl);
            const port = parsed.port || (parsed.protocol === 'https:' ? 443 : 80);
            const path = normalizePath(parsed.pathname, parsed.search, parsed.hash);
            const deviceType = inferDeviceType(port);

            // Health check: verify final URL is accessible
            try {
                await axios.get(finalUrl, { timeout: 3000, validateStatus: () => true });
                await updateDeviceNetworkInfo(deviceName, {
                    ip: parsed.hostname,
                    port,
                    path,
                    finalUrl,
                    deviceType,
                });
                console.log(`Resolved (ID: ${deviceName}) -> ${finalUrl}`);
                return { url: finalUrl, deviceId: deviceName, type: deviceType };
            } catch (healthErr) {
                console.warn(`Device (ID: ${deviceName}) URL unreachable: ${finalUrl} - ${healthErr.message}`);
                await updateDeviceNetworkInfo(deviceName, {
                    ip: parsed.hostname,
                    port,
                    path,
                    finalUrl,
                    deviceType,
                });
                return { url: finalUrl, deviceId: deviceName, type: deviceType, status: 'offline' };
            }
        }
    }

    console.warn(`Could not get redirect for (ID: ${deviceName}). Status: ${lastStatus} (${lastUrl})`);
    return null;
};

app.get('/api/devices', (req, res) => {
    const { branch } = req.query;
    let query = "SELECT * FROM devices";
    let params = [];
    
    if (branch) {
        query += " WHERE branch = ?";
        params.push(branch);
    }
    
    db.all(query, params, (err, rows) => {
        if (err) {
            res.status(500).json({ "error": err.message });
            return;
        }
        res.json({
            "message": "success",
            "data": rows
        });
    });
});

app.get('/api/branches', (req, res) => {
    const branches = BRANCHES_DATA.map((branch) => ({
        code: branch.code,
        prefixed: ensurePrefixed(branch.code),
        name: branch.name,
        address: branch.address,
    }));
    res.json({ message: 'success', data: branches });
});

app.post('/api/devices', (req, res) => {
    const { name, branch } = req.body;
    
    if (!branch) {
        res.status(400).json({ "error": "Branch là bắt buộc." });
        return;
    }
    
    // Check for duplicate within the same branch
    db.get(`SELECT id FROM devices WHERE name = ? AND branch = ?`, [name, branch], (err, row) => {
        if (err) {
            res.status(500).json({ "error": err.message });
            return;
        }
        if (row) {
            res.status(409).json({ "error": `Phòng "${name}" đã tồn tại trong chi nhánh này. Vui lòng chọn tên khác.` });
            return;
        }
        db.run(`INSERT INTO devices (name, branch) VALUES (?, ?)`, [name, branch], function(err) {
            if (err) {
                res.status(400).json({ "error": err.message });
                return;
            }
            res.json({
                "message": "success",
                "data": { id: this.lastID, name, branch }
            });
        });
    });
});

app.put('/api/devices/:id', (req, res) => {
    const { name } = req.body;
    const { id } = req.params;
    // Check for duplicate (excluding current device)
    db.get(`SELECT id FROM devices WHERE name = ? AND id != ?`, [name, id], (err, row) => {
        if (err) {
            res.status(500).json({ "error": err.message });
            return;
        }
        if (row) {
            res.status(409).json({ "error": `Phòng "${name}" đã tồn tại. Vui lòng chọn tên khác.` });
            return;
        }
        db.run(`UPDATE devices SET name = ? WHERE id = ?`, [name, id], function(err) {
            if (err) {
                res.status(400).json({ "error": err.message });
                return;
            }
            res.json({ message: "success", data: { id, name } });
        });
    });
});

app.delete('/api/devices/:id', (req, res) => {
    db.run(`DELETE FROM devices WHERE id = ?`, req.params.id, function(err) {
        if (err) {
            res.status(400).json({ "error": err.message });
            return;
        }
        res.json({ message: "deleted", changes: this.changes });
    });
});

// Scan devices: always hit network resolver
app.get('/scan/network', async (req, res) => {
    console.log('Received request to /scan/network');
    const { branch } = req.query;
    
    if (!branch) {
        return res.status(400).json({ error: 'Branch parameter is required' });
    }
    
    try {
        db.all("SELECT * FROM devices WHERE branch = ?", [branch], async (err, rows) => {
            if (err) return res.status(500).json({ error: err.message });
            const resolvedDevices = (
                await Promise.all(
                    rows.map((device) => resolveViaNetwork(device?.name, branch).catch((e) => {
                        console.error(`Network resolve failed for ${device?.name}:`, e.message);
                        return null;
                    }))
                )
            ).filter(Boolean);
            console.log('Network-only scan complete. Devices:', resolvedDevices.length);
            res.json(resolvedDevices);
        });
    } catch (error) {
        console.error('Error during network-only scan:', error);
        res.status(500).json({ error: 'Failed to scan devices (network-only)' });
    }
});

// Scan devices: cache-first local check
app.get('/scan/local', async (req, res) => {
    console.log('Received request to /scan/local');
    const { branch } = req.query;

    if (!branch) {
        return res.status(400).json({ error: 'Branch parameter is required' });
    }

    try {
        db.all("SELECT * FROM devices WHERE branch = ?", [branch], async (err, rows) => {
            if (err) return res.status(500).json({ error: err.message });

            const resolvedDevices = await Promise.all(
                rows.map(async (device) => {
                    if (!device.last_url) {
                        return { url: null, deviceId: device.name, type: device.device_type || 'unknown', status: 'cached_url_unavailable' };
                    }
                    try {
                        // Quick health check on the cached URL
                        await axios.get(device.last_url, { timeout: 2000, validateStatus: () => true });
                        return { url: device.last_url, deviceId: device.name, type: device.device_type };
                    } catch (healthErr) {
                        return { url: device.last_url, deviceId: device.name, type: device.device_type, status: 'offline' };
                    }
                })
            );
            console.log('Local-first scan complete. Devices:', resolvedDevices.length);
            res.json(resolvedDevices);
        });
    } catch (error) {
        console.error('Error during local-first scan:', error);
        res.status(500).json({ error: 'Failed to scan devices (local-first)' });
    }
});

// Global error handler - must be last
app.use((err, req, res, next) => {
    console.error('Unhandled error:', err);
    res.status(500).json({ 
        error: 'Internal server error',
        message: err.message 
    });
});

// Start server after database is initialized
const startServer = async () => {
    try {
        console.log('Initializing database...');
        await initDatabase();
        console.log('✓ Database initialized');

        const server = app.listen(port, () => {
            console.log(`✓ Backend server listening at http://localhost:${port}`);
        });

        // Handle server errors
        server.on('error', (err) => {
            console.error('Server error:', err);
            process.exit(1);
        });
    } catch (err) {
        console.error('Failed to start server:', err);
        process.exit(1);
    }
};

startServer();